弾の当たり判定表示（デバック）の追加、無駄なデバック用のBulletBuffer の削除

+++ Assets/Scripts/Bullets/WarpBulletJob.cs
@@ -131,7 +131,7 @@ public struct WarpBulletJob : IJobParallelFor
 
     private float GetAngleRad(float x, float y)
     {
-        double rad = math.atan2(y, x);
+        float rad = math.atan2(y, x);
         if (rad < 0) rad += 2 * math.PI;
         return (float)rad;
     }
